fastapi_api/requirements.txt
fastapi
uvicorn[standard]
requests
boto3
botocore
sqlalchemy
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
pydantic
email-validator
psycopg2-binary
